Added additional util functions, Added draw debug for collider2d, Fixed wrong scaling of Collider2d when having a parent, Fixed wrong scaling of Sprite when having a parent, Added seperate draw function for Sprite, Fixed getting texture by id in Texture2d, Fixed formatting, Fixed input being limited to 256 keys, Added ability to toggle debug window, Added chunking system, Added map edit window, Added structures for objects and chunks, Added simple loading and saving of chunks

+++ libs/engine/io/Input.c
@@ -2,42 +2,42 @@
 
 #include "raylib.h"
 
-bool Input_IsKeyPressed(uint8_t key)
+bool Input_IsKeyPressed(uint16_t key)
 {
     return IsKeyPressed(key);
 }
 
-bool Input_IsKeyDown(uint8_t key)
+bool Input_IsKeyDown(uint16_t key)
 {
     return IsKeyDown(key);
 }
 
-bool Input_IsKeyReleased(uint8_t key)
+bool Input_IsKeyReleased(uint16_t key)
 {
     return IsKeyReleased(key);
 }
 
-bool Input_IsKeyUp(uint8_t key)
+bool Input_IsKeyUp(uint16_t key)
 {
     return IsKeyUp(key);
 }
 
-bool Input_IsMouseButtonPressed(uint8_t button)
+bool Input_IsMouseButtonPressed(uint16_t button)
 {
     return IsMouseButtonPressed(button);
 }
 
-bool Input_IsMouseButtonDown(uint8_t button)
+bool Input_IsMouseButtonDown(uint16_t button)
 {
     return IsMouseButtonDown(button);
 }
 
-bool Input_IsMouseButtonReleased(uint8_t button)
+bool Input_IsMouseButtonReleased(uint16_t button)
 {
     return IsMouseButtonReleased(button);
 }
 
-bool Input_IsMouseButtonUp(uint8_t button)
+bool Input_IsMouseButtonUp(uint16_t button)
 {
     return IsMouseButtonUp(button);
 }
